Aäron Vandeputte's changes to add Sulfur to RMG database
Aaron emailed us a zip file of his database at the end of
the summer. I have tried to locate the closest commit to
where he diverged, and branched from there, before pasting
over with his version. I have also had to play with line
endings and file permissions to try to minimize the diff.
I then rebased it onto a new branch that had nothing but
line endings fixed, so that the line-ending-only changes
would be removed, before cherry-picking it onto a third
branch all of its own.

I hope it works!

Signed-off-by: Richard West <rwest@mit.edu>